+++ mhpasswordmanager-password-service/src/main/java/br/com/marcoshssilva/mhpasswordmanager/passwordservice/domain/services/crypt/CryptServiceConfig.java
@@ -1,5 +1,7 @@
 package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt;
 
+import br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.impl.AESCryptServiceImpl;
+import br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.impl.RSACryptServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
