+++ mhpasswordmanager-password-service/src/main/java/br/com/marcoshssilva/mhpasswordmanager/passwordservice/domain/services/crypt/exceptions/EncryptionException.java
@@ -1,4 +1,4 @@
-package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt;
+package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.exceptions;
 
 public class EncryptionException extends RuntimeException{
     public EncryptionException(String message, Throwable cause) {
