+++ mhpasswordmanager-password-service/src/main/java/br/com/marcoshssilva/mhpasswordmanager/passwordservice/domain/services/crypt/impl/AESCryptServiceImpl.java
@@ -1,5 +1,8 @@
-package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt;
+package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.impl;
 
+import br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.CryptService;
+import br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.exceptions.DecryptionException;
+import br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.exceptions.EncryptionException;
 import lombok.RequiredArgsConstructor;
 
 import javax.crypto.Cipher;
