+++ mhpasswordmanager-password-service/src/test/java/br/com/marcoshssilva/mhpasswordmanager/passwordservice/domain/services/crypt/exceptions/EncryptionExceptionTests.java
@@ -1,5 +1,6 @@
-package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt;
+package br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.exceptions;
 
+import br.com.marcoshssilva.mhpasswordmanager.passwordservice.domain.services.crypt.exceptions.EncryptionException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
